adicionado useHistory nos repositorios para retornar para home quando repositoriesName for nulo

+++ src/pages/Repositories/Repositories.js
@@ -1,16 +1,25 @@
 import React, { useState, useEffect } from "react";
+import { useHistory } from "react-router-dom";
 
 import * as S from './styled'
 
 function Repositories () {
 
+  const history = useHistory();
   const [ repositories, setRepositories ] = useState([]);
   useEffect(() => {
     let repositoriesName = localStorage.getItem('repositoriesName');
-    repositoriesName = JSON.parse(repositoriesName);
-    setRepositories(repositoriesName);
-    localStorage.clear();
+
+    if(repositoriesName !== null){
+      repositoriesName = JSON.parse(repositoriesName);
+      setRepositories(repositoriesName);
+      localStorage.clear();
+    }else{
+      history.push("/");
+    }
+    
   }, []);
+
   return (
     <S.Container>
       <S.Title>Repositories</S.Title>
